Arreglado el store y reducer, cambie extenciones de componentes de tsx a jsx
