+++ client/src/Redux/store.js
@@ -1,6 +1,15 @@
-import { createStore, applyMiddleware } from "redux"
-import { composeWithDevTools } from "redux-devtools-extension"
-import thunk from "redux-thunk"
-import rootReducer from "./reducer"
+import { createStore, applyMiddleware, compose } from "redux";
+import { rootReducer } from "./reducer";
+import thunk from "redux-thunk";
 
-export const store = createStore(rootReducer, composeWithDevTools(applyMiddleware(thunk)))
+const composeEnhancers =
+    (typeof window !== "undefined" &&
+        window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) ||
+    compose;
+
+const store = createStore(
+    rootReducer,
+    composeEnhancers(applyMiddleware(thunk))
+);
+
+export default store;
